fastapi-app/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-grpc>=1.20
opentelemetry-instrumentation-fastapi>=0.41b0
python-json-logger>=2.0
